robot_client/requirements.txt
# TonyPi Robot Client Requirements
# Install these on your Raspberry Pi with TonyPi robot

# Core MQTT communication
paho-mqtt>=2.0.0
psutil>=5.9.0

# Hardware SDK dependencies (required on TonyPi Raspberry Pi)
pyserial>=3.5           # Serial communication with STM32 board
smbus2>=0.4.2           # I2C communication for ultrasonic sensor

# Camera streaming (optional, for video feed)
opencv-python>=4.8.0    # Camera capture and MJPEG encoding
numpy>=1.24.0           # Image processing

# Configuration
pyyaml>=6.0             # YAML config file parsing

# Optional: For advanced features
# RPi.GPIO>=0.7.1       # GPIO control (pre-installed on Raspberry Pi OS)
# gpiozero>=1.6.2       # High-level GPIO interface
